move into PackedSequence
